pyproject.toml
[tool.poetry]
name = "sway-remap"
version = "0.1.0"
description = ""
authors = ["Kay Gosho <ketsume0211@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-uinput = "^0.11.2"
evdev = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
